Pulled create_name and edit_name out of name_controller.rb and made a new file in the subdirectory.  Refactred them, but mostly only superficially.  I tried my best to improve the understandability of these actions, but only helped marginally.  Not sure why it's so difficult.
